Split utils file into multiple files
Splitting the utils file into multiple files:

- api: Contains all methods for API requests. This includes requests
against the AWS and Google API and requests against the Kubernetes API
server.

- constants: Contains now also the keys for items whiche are saved into
the localStorage.

- storage: Contains all operations whiche are executed against
localStorage.

- helpers: Contains all other helper functions.
