+++ src/components/kubernetes/Conditions.tsx
@@ -7,7 +7,7 @@ import {
 import React from 'react';
 
 import { TCondition } from '../../declarations';
-import { getProperty, timeDifference } from '../../utils';
+import { getProperty, timeDifference } from '../../utils/helpers';
 import IonCardEqualHeight from '../misc/IonCardEqualHeight';
 
 interface IPodConditionsProps {
