+++ src/components/kubernetes/EditItem.tsx
@@ -15,8 +15,8 @@ import { close, create } from 'ionicons/icons';
 import yaml from 'js-yaml';
 import React, { useContext, useState } from 'react';
 
-import { AppContext } from '../../context';
 import { IContext, TActivator } from '../../declarations';
+import { AppContext } from '../../utils/context';
 import Editor from '../misc/Editor';
 
 interface IEditItemProps {
