+++ src/components/kubernetes/details/Deployment.tsx
@@ -8,7 +8,7 @@ import { V1Deployment, V1DeploymentStrategy } from '@kubernetes/client-node'
 import React from 'react';
 import { RouteComponentProps } from 'react-router';
 
-import { labelSelector } from '../../../utils';
+import { labelSelector } from '../../../utils/helpers';
 import Conditions from '../Conditions';
 import Configuration from '../Configuration';
 import List from '../List';
